Enable pre-commit hooks
Enable pre-commit hooks:  standard hooks, shellcheck, and ksconf check) with
resulting trivial tweaks:

- Various minor shell script fixes based on feedback from shellcheck
- Python 3.x syntax compatibility fix (exception 'as' vs comma)
- Consolidate duplicate .conf stanza

+++ fetch_deps.sh
@@ -3,4 +3,4 @@
 LIB=bin/lib
 
 python2.7 -m pip install --upgrade jmespath --target=$LIB
-rm -rf $LIB/bin $LIB/*.dist-info
+rm -rf ${LIB:?}/bin $LIB/*.dist-info
